requirements.txt
Flask==3.0.0
flask-cors==4.0.0
torch==2.0.1
transformers==4.44.0
pypdf2==3.0.1
python-docx
pytesseract
Pillow
openai
accelerate
numpy